Restructured the whole Projekt Refactoring Added a Parser for Reponses Added Exceptions Added Methods to get the current Courses in the current semester Added Methods to fetch all News for a specific Course
