+++ src/main/java/src/de/oelrichsgarcia/studipTelegramBot/studip/BotException.java
@@ -1,0 +1,14 @@
+package de.oelrichsgarcia.studipTelegramBot.studip;
+
+/**
+ * The interface Bot exception.
+ */
+public interface BotException {
+
+    /**
+     * Gets the errormessage.
+     *
+     * @return the message
+     */
+    String getErrorMessage();
+}
