+++ src/main/java/src/de/oelrichsgarcia/studipTelegramBot/studip/NotLoggedInException.java
@@ -1,0 +1,9 @@
+package de.oelrichsgarcia.studipTelegramBot.studip;
+
+public class NotLoggedInException extends GeneralBotException {
+
+
+    public NotLoggedInException() {
+        super("YOu have to login first");
+    }
+}
