+++ src/main/java/src/de/oelrichsgarcia/studipTelegramBot/studip/api/request/APIRequest.java
@@ -1,0 +1,18 @@
+package de.oelrichsgarcia.studipTelegramBot.studip.api.request;
+
+import java.io.IOException;
+
+/**
+ * The interface Api request.
+ */
+interface APIRequest {
+
+    /**
+     * Gets response.
+     *
+     * @return the response
+     * @throws IOException  the io exception
+     * @throws APIException the api exception
+     */
+    RequestResponse getResponse() throws IOException, APIException;
+}
