--- src/studip/api/request/APIException.java
@@ -1,28 +1,0 @@
-package studip.api.request;
-
-public class APIException extends Exception {
-
-    private int errorCode;
-    private String errorMessage;
-
-    public APIException(int errorCode, String errorMessage) {
-        this.errorCode = errorCode;
-        this.errorMessage = errorMessage;
-    }
-
-    public int getErrorCode() {
-        return errorCode;
-    }
-
-    public void setErrorCode(int errorCode) {
-        this.errorCode = errorCode;
-    }
-
-    public String getErrorMessage() {
-        return errorMessage;
-    }
-
-    public void setErrorMessage(String errorMessage) {
-        this.errorMessage = errorMessage;
-    }
-}
