--- src/studip/api/request/APIRequest.java
@@ -1,8 +1,0 @@
-package studip.api.request;
-
-import java.io.IOException;
-
-interface APIRequest {
-
-    RequestResponse getResponse() throws IOException, APIException;
-}
