--- src/studip/api/request/LoginException.java
@@ -1,8 +1,0 @@
-package studip.api.request;
-
-public class LoginException extends APIException {
-
-    public LoginException(int errorCode, String errorMessage) {
-        super(errorCode, errorMessage);
-    }
-}
