--- src/studip/api/request/RequestException.java
@@ -1,9 +1,0 @@
-package studip.api.request;
-
-public class RequestException extends APIException {
-
-
-    public RequestException(int errorCode, String errorMessage) {
-        super(errorCode, errorMessage);
-    }
-}
